sub comm

+++ src/circle.cpp
@@ -2,7 +2,7 @@
 #include <iostream>
 #include <numbers>
 
-Circle::Circle(Point center, std::string name, int radius) : Shape(center, name), radius(radius){}
+Circle::Circle(Point center, std::string name, double radius) : Shape(center, name), radius(radius){}
 
 double Circle::area() const{
   return std::numbers::pi * radius * radius;
