+++ src/rectangle.cpp
@@ -1,7 +1,7 @@
 #include "rectangle.hpp"
 #include <iostream>
 
-Rectangle::Rectangle(Point center, std::string name, int width, int height): Shape(center, name), width(width), height(height){}
+Rectangle::Rectangle(Point center, std::string name, double width, double height): Shape(center, name), width(width), height(height){}
 
 double Rectangle::area() const{
   return width*height;
