+++ src/rectangle.hpp
@@ -4,7 +4,7 @@
 
 class Rectangle : public Shape{
 public:
-  Rectangle(Point center, std::string name, int width, int height);
+  Rectangle(Point center, std::string name, double width, double height);
 
   double area() const override;
   void draw(std::ostream& out) const override;
@@ -14,8 +14,8 @@ protected:
   Rectangle(const Rectangle& other) = default;
 
 private:
-  int width;
-  int height; 
+  double width;
+  double height; 
 
 
 };
