+++ src/square.cpp
@@ -1,6 +1,6 @@
 #include "square.hpp"
 
-Square::Square(Point center, std::string name, int side): Rectangle(center, name, side, side){}
+Square::Square(Point center, std::string name, double side): Rectangle(center, name, side, side){}
 
 Square *Square::clone() const{
   return new Square(*this);
