+++ src/square.hpp
@@ -4,7 +4,7 @@
 
 class Square : public Rectangle{
 public:
-Square(Point center, std::string name, int side);
+Square(Point center, std::string name, double side);
 
   Square* clone() const override;
